requirements.txt
Django==4.2.20
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.7.0
djoser==2.2.0  # Critical update from 2.1.0
social-auth-app-django==5.4.3  # Now compatible with djoser 2.2.0
social-auth-core==4.5.6
python-decouple==3.8
asgiref==3.8.1
pillow==11.2.1
djangorestframework-simplejwt==4.8.0
typing-extensions==4.9.0
